Split reducer common helpers
